Switch day01a to int64

+++ src/day01a.f90
@@ -1,5 +1,6 @@
 !> Solution for https://adventofcode.com/2023/day/1 part a
 module day01a
+    use iso_fortran_env, only : int64
     use util, only : readinputfile_asstringarray, code_0, code_9
     implicit none
     private
@@ -14,7 +15,7 @@ contains
         implicit none
 
         character(len=*), intent(in)  :: lines(:)
-        integer                       :: calibration_value
+        integer(int64)                :: calibration_value
         character(len=:), allocatable :: line
         integer                       :: i, j, digit
 
@@ -42,12 +43,12 @@ contains
 
     end function sumnumbers
 
-    integer function solve(filename)
+    integer(int64) function solve(filename)
         implicit none
 
         character(len=*), intent(in)  :: filename
         character(len=:), allocatable :: lines(:)
-        integer                       :: calibration_value
+        integer(int64)                :: calibration_value
 
         lines = readinputfile_asstringarray(filename, maxlinelength)
         calibration_value = sumnumbers(lines)
